Fixed a race condition with callback queues which required an interface-breaking change to CallbackQueue/CallbackQueueInterface
This change adds a mechanism for removing all callbacks associated with a specific set of user data (a uint64).  The specific bug this fixes is the case where:

 * callback added
 * object the callback referenced is destroyed (and the subscription/timer/etc. is removed correctly)
 * callOne()/callAvailable() is called -- callback still exists in the queue, and is attempted to be called

Most of the specific callbacks being used had some reason that this would not happen often in practice.  subscriber callbacks were the only case that happened deterministically.  The rest showed up as nondeterministic crashes from timer/subscription/service callbacks.