Re-implemented WITH-FIELDS
The new implementation handles type declarations and uses them for optimizing
message slot access. Also, added new function MSG-SLOT-VALUE that is similar to
slot value but intended to be used for message access. It ignores the package
of the symbol indicating the sot to be accessed and uses the package of the
message class instead.