rospy compatibility for Python 2.7/Natty.
This is a fairly major change that removes caching
of the xmlrpclib.ServerProxy objects.  In Py27, these
no longer appear to be threadsafe in use, so the
best policy is to not reuse them and taken the very
minor performance hit.

There is a small optimization introduced in tcpros_base
to use a lock around checking the tcp keepalive parameter.
This lock ensures that only one outgoing call occurs.