Removed logic to notice that sudo is in use and drop permissions.  Instead of determining home by examining $HOME, we look up the password entry based on effective UID and extract pw_dir.
The result is that when rospack runs as root (sudo or otherwise), it will
default to writing the cache into root's home directory, instead of
dropping perms and reusing the non-root user's cache.

This is hopefully a final fix for #2884.

rosstack still has sudo-related problems, but they're not easily fixed, as
it has a totally different code path for determining the cache location
(yar!).