Added shell functions for fish shell
This commit adds a set of functions for the fish shell that mirror those
provided by rosbash.
Autocompletions are not provided here.
rospd differs from that provided by rosbash in the same way that fish's
pd differ's from bash's.
This commit works with the 2.0 branch of fish under current development
on github, and it is incompatible with the legacy version in Debian.
Hopefully with 2.0 finally reached it'll become the default soon.