Adding support for fish shell
This adds a rosfish alongside rosbash and the others to provide shell
functions such as roscd and appropriate tab completions.  Due to
differing syntax fish users will have to set environemnt variables
manually that are set by setup.sh for bash and zsh but once that is done
they should be able to use all the same facilities as other supported
shells.